+++ src/main/java/com/example/kuyou/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
 
     public Map<String, Object> getOtherinfo(long id);
     public int resourceRequest(String resource_post_kind,long resource_id,long u_id );
-    public int postTravelstate(String position,int u_id);
+    public int postTravelstate(String position,long u_id);
 
     public List<User> getFanslist(long u_id,long target_id);
     public List<User> getFollowslist(long u_id,long target_id);
