+++ src/main/java/com/example/kuyou/service/impl/FileUploadServiceImpl.java
@@ -22,7 +22,7 @@ public class FileUploadServiceImpl implements FileUploadService {
     UserDao userDao;
 
     //用于更新label表，video表，join-label表
-    public String uploadVideo(int v_creator,String position,int m_id,String label,String v_description,String v_date,String v_content,String v_cover){
+    public String uploadVideo(long v_creator,String position,int m_id,String label,String v_description,String v_date,String v_content,String v_cover){
 
         //旅行地址解析
         String str[]=position.split("·");
@@ -83,12 +83,12 @@ public class FileUploadServiceImpl implements FileUploadService {
 
     public List<Labelusenum> searchLabel(String labels){return fileUploadDao.searchLabel(labels);}
 
-    public void alterBackcover(int u_id, String bg_img){
+    public void alterBackcover(long u_id, String bg_img){
         Map<String,Object> group=new HashMap<String, Object>();
         group.put("u_id",u_id);group.put("bg_img",bg_img);
         fileUploadDao.alterBackcover(group);}
 
-    public void alterHeadcover(int u_id, String head){
+    public void alterHeadcover(long u_id, String head){
         Map<String,Object> group=new HashMap<String, Object>();
         group.put("u_id",u_id);group.put("bg_img",head);
         fileUploadDao.alterBackcover(group);}
